[#116] refactor: 회원 쿼리 패키지 구조 수정
